[DEV] Handle Relay groups in solenoid Manager

+++ embedded/main/Utils.hpp
@@ -15,5 +15,5 @@ namespace Utils
 String GetSubStr(const String& p_rawMsg, int p_startId, int p_endId, bool p_verbose = false);
 int32_t GetSmoothedRSSI(int32_t newRSSI);
 
-uint8_t scaleTo99(uint16_t value);
+uint16_t scaleTo99(uint16_t value);
 } // namespace Utils
